requirements.txt
openai-whisper
openai
pyttsx3